ch(implement-feedback): move all red-flags controllers
- move all red-flags controllers into into one file
- delete the red-flags controller folder
- configure body-parser in server/index.js file
- add a command to .eslintrc file

[Start #162411545]
