+++ .eslintrc.json
@@ -7,6 +7,7 @@
   },
   "rules": {
     "linebreak-style": 0,
-    "consistent-return": 0
+    "consistent-return": 0,
+    "import/prefer-default-export": "off"
   }
 }
