--- server/controllers/incidents/red-flag/getRedFlags.js
@@ -1,10 +1,0 @@
-import db from '../../../models/red-flag';
-
-const getRedFlags = (req, res) => {
-  res.json({
-    status: 200,
-    data: db, // Get all the red-flags from the DB(data structure)
-  });
-};
-
-export default getRedFlags;
