+++ server/controllers/index.js
@@ -1,9 +1,11 @@
-import getRedFlags from './incidents/red-flag/getRedFlags';
-import getSpecificRedFlag from './incidents/red-flag/getSpecificRedFlag';
-import createRedFlag from './incidents/red-flag/createRedFlag';
-import editLocation from './incidents/red-flag/editRedFlagLocation';
-import editComment from './incidents/red-flag/editRedFlagComment';
-import deleteRedFlag from './incidents/red-flag/deleteRedFlag';
+import {
+  createRedFlag,
+  getRedFlags,
+  getSpecificRedFlag,
+  editLocation, editComment,
+  deleteRedFlag,
+} from './incidents/red-flags';
+
 
 export default {
   getRedFlags,
