+++ server/middleware/index.js
@@ -4,20 +4,23 @@ const middlewareObj = {};
 middlewareObj.checkUserInput = (req, res, next) => {
 //  checking if the user's input is valid
   if (!req.body.type || req.body.type.toLowerCase() !== 'red-flag') {
-    return res.json({
+    res.json({
       status: 400,
       error: 'Incidient type must be \'red-flag\' ',
     });
+    return;
   } if (!req.body.location || req.body.location === '') {
-    return res.json({
+    res.json({
       status: 400,
       error: 'Location is required',
     });
+    return;
   } if (!req.body.comment || req.body.comment === '') {
-    return res.json({
+    res.json({
       status: 400,
       error: 'comment is required',
     });
+    return;
   }
 
   next();
